Only expose function for showing the equilibrium computation monitor
